refactor(handoff): extract ExecToPlanExecutor.js into modular structure
Part of SD-LEO-REFACTOR-EXECTOPLAN-001 (Child 6 of orchestrator SD-LEO-REFACTOR-LARGE-FILES-002)

Refactors 1,271 LOC monolithic ExecToPlanExecutor.js into focused modules:
- gates/prerequisite-check.js - PLAN-TO-EXEC prerequisite validation
- gates/test-evidence-auto-capture.js - LEO v4.4.2 auto-capture
- gates/sub-agent-orchestration.js - Sub-agent orchestration
- gates/mandatory-testing-validation.js - Mandatory testing gate
- gates/bmad-validation.js - BMAD validation
- gates/gate2-implementation-fidelity.js - Implementation fidelity
- gates/rca-gate.js - RCA validation
- gates/human-verification-gate.js - Human verification (LEO v4.4.0)
- gates/index.js - Gate exports
- state-transitions.js - User story/PRD/SD transitions
- test-evidence.js - Test evidence validation
- git-verification.js - Git commit verification, auto-shipping
- russian-judge.js - AI quality assessment
- remediation.js - Gate remediation messages
- index.js - Main executor class with 21 exports

Testing: Dynamic import verified, ESLint clean
Total: 1,572 LOC in modular structure, original file reduced to 14 LOC re-export
