+++ scripts/modules/handoff/executors/ExecToPlanExecutor.js
@@ -1,1271 +1,14 @@
 /**
  * ExecToPlanExecutor - Executes EXEC → PLAN handoffs
- * Part of LEO Protocol Unified Handoff System refactor
+ * Part of LEO Protocol Unified Handoff System
  *
- * Validates that EXEC phase implementation is complete and ready for PLAN verification.
+ * This file re-exports from the modular exec-to-plan directory structure.
+ * Refactored as part of SD-LEO-REFACTOR-EXECTOPLAN-001.
  *
- * SD-LEO-ID-NORMALIZE-001: Uses SD ID normalizer for all update operations.
+ * @see ./exec-to-plan/index.js for main implementation
  */
 
-import BaseExecutor from './BaseExecutor.js';
-import { normalizeSDId } from '../../sd-id-normalizer.js';
-
-// External validators (will be lazy loaded)
-let validateBMADForExecToPlan;
-let validateGate2ExecToPlan;
-let orchestrate;
-let getValidationRequirements;
-let mapE2ETestsToUserStories;
-let validateE2ECoverage;
-let autoValidateUserStories;
-let autoCompleteDeliverables;
-let checkDeliverablesNeedCompletion;
-// LEO v4.3.4: Unified test evidence functions
-let getStoryTestCoverage;
-
-export class ExecToPlanExecutor extends BaseExecutor {
-  constructor(dependencies = {}) {
-    super(dependencies);
-    this.validators = dependencies.validators || {};
-  }
-
-  get handoffType() {
-    return 'EXEC-TO-PLAN';
-  }
-
-  async setup(sdId, sd, _options) {
-    await this._loadValidators();
-
-    // Check sd_type for documentation-only SDs
-    const validationReqs = getValidationRequirements(sd);
-    console.log(`\n📋 SD Type: ${sd.sd_type || 'feature (default)'}`);
-
-    if (validationReqs.skipCodeValidation) {
-      console.log('   ✅ DOCUMENTATION-ONLY SD DETECTED');
-      console.log('   → TESTING/GITHUB validation will be SKIPPED');
-      console.log(`   → Reason: ${validationReqs.reason}`);
-    }
-
-    return null;
-  }
-
-  getRequiredGates(_sd, _options) {
-    const gates = [];
-
-    // ROOT CAUSE FIX: Prerequisite handoff validation
-    // SD-VISION-V2-009 identified gap: EXEC-TO-PLAN could proceed without accepted PLAN-TO-EXEC
-    // This gate ensures the LEO Protocol handoff chain is enforced sequentially
-    gates.push({
-      name: 'PREREQUISITE_HANDOFF_CHECK',
-      validator: async (ctx) => {
-        console.log('\n🔐 PREREQUISITE CHECK: PLAN-TO-EXEC Handoff Required');
-        console.log('-'.repeat(50));
-
-        // Query for PLAN-TO-EXEC handoff (any status - to detect blocked handoffs)
-        // Use UUID (ctx.sd.id) not legacy_id (ctx.sdId) - handoffs are stored by UUID
-        const sdUuid = ctx.sd?.id || ctx.sdId;
-        const { data: allHandoffs, error } = await this.supabase
-          .from('sd_phase_handoffs')
-          .select('id, status, created_at, validation_score, rejection_reason')
-          .eq('sd_id', sdUuid)
-          .eq('handoff_type', 'PLAN-TO-EXEC')
-          .order('created_at', { ascending: false })
-          .limit(5);
-
-        if (error) {
-          console.log(`   ⚠️  Error checking prerequisite: ${error.message}`);
-          return {
-            passed: false,
-            score: 0,
-            max_score: 100,
-            issues: [`Failed to verify PLAN-TO-EXEC prerequisite: ${error.message}`],
-            warnings: []
-          };
-        }
-
-        // ROOT CAUSE FIX (2026-01-01): Check for ACCEPTED handoff FIRST before checking blocked
-        // An accepted handoff takes precedence over older blocked handoffs
-        // This fixes SD-VS-CHAIRMAN-SETTINGS-001 where accepted handoff (86%) was being
-        // ignored because an older blocked handoff (0%) existed
-        const acceptedHandoffs = allHandoffs?.filter(h => h.status === 'accepted') || [];
-        const blockedHandoffs = allHandoffs?.filter(h => h.status === 'blocked') || [];
-
-        // If we have at least one accepted handoff, the prerequisite is satisfied
-        if (acceptedHandoffs.length > 0) {
-          const latestAccepted = acceptedHandoffs[0]; // Already sorted by created_at desc
-          console.log('   ✅ Prerequisite satisfied: PLAN-TO-EXEC handoff found');
-          console.log(`      Handoff ID: ${latestAccepted.id.slice(0, 8)}...`);
-          console.log(`      Status: ${latestAccepted.status}`);
-          console.log(`      Score: ${latestAccepted.validation_score}`);
-          console.log(`      Date: ${new Date(latestAccepted.created_at).toLocaleString()}`);
-
-          // Warn if there are also blocked handoffs (but don't fail)
-          if (blockedHandoffs.length > 0) {
-            console.log(`   ⚠️  Note: ${blockedHandoffs.length} earlier blocked handoff(s) exist (ignored - accepted handoff takes precedence)`);
-          }
-
-          return {
-            passed: true,
-            score: 100,
-            max_score: 100,
-            issues: [],
-            warnings: blockedHandoffs.length > 0 ? [`${blockedHandoffs.length} blocked handoff(s) exist from earlier attempts`] : []
-          };
-        }
-
-        // No accepted handoff - check if there are blocked handoffs to provide guidance
-        if (blockedHandoffs.length > 0) {
-          const latestBlocked = blockedHandoffs[0];
-          console.log('   ⚠️  BLOCKED PLAN-TO-EXEC handoff found (no accepted handoff exists)');
-          console.log(`      ID: ${latestBlocked.id.slice(0, 8)}...`);
-          console.log(`      Score: ${latestBlocked.validation_score}% (required: 85%)`);
-          console.log(`      Reason: ${latestBlocked.rejection_reason || 'Below threshold'}`);
-          console.log('\n   REMEDIATION:');
-          console.log('   1. Address validation failures to raise score to 85%+');
-          console.log(`   2. Use: SELECT * FROM retry_blocked_handoff('${latestBlocked.id}', <new_score>);`);
-          console.log('   3. Or create new handoff after fixing issues');
-
-          return {
-            passed: false,
-            score: 0,
-            max_score: 100,
-            issues: [`BLOCKED: PLAN-TO-EXEC handoff blocked with score ${latestBlocked.validation_score}%`],
-            warnings: [],
-            remediation: 'Fix validation issues and retry blocked handoff or create new one'
-          };
-        }
-
-        // No handoffs at all
-        // SD-LEARN-010:US-002: ERR_CHAIN_INCOMPLETE error code for missing predecessor handoffs
-        const planToExecHandoff = acceptedHandoffs;
-        if (!planToExecHandoff || planToExecHandoff.length === 0) {
-          console.log('   ❌ ERR_CHAIN_INCOMPLETE: Missing PLAN-TO-EXEC handoff');
-          console.log('   ⚠️  LEO Protocol requires PLAN-TO-EXEC before EXEC-TO-PLAN');
-          console.log('\n   REMEDIATION:');
-          console.log('   1. Complete PLAN phase prerequisites (PRD, user stories, design analysis)');
-          console.log('   2. Run: node scripts/handoff.js execute PLAN-TO-EXEC <SD-ID>');
-          console.log('   3. Address any validation failures');
-          console.log('   4. Retry EXEC-TO-PLAN after PLAN-TO-EXEC is accepted');
-
-          return {
-            passed: false,
-            score: 0,
-            max_score: 100,
-            issues: ['ERR_CHAIN_INCOMPLETE: Missing PLAN-TO-EXEC handoff - complete prerequisite before EXEC-TO-PLAN'],
-            warnings: [],
-            remediation: 'Complete PLAN-TO-EXEC handoff before attempting EXEC-TO-PLAN'
-          };
-        }
-
-        const handoff = planToExecHandoff[0];
-        console.log(`   ✅ PLAN-TO-EXEC handoff found: ${handoff.id.slice(0, 8)}...`);
-        console.log(`      Status: ${handoff.status}`);
-        console.log(`      Score: ${handoff.validation_score}%`);
-        console.log(`      Date: ${new Date(handoff.created_at).toLocaleString()}`);
-
-        // Store for later reference
-        ctx._planToExecHandoff = handoff;
-
-        return {
-          passed: true,
-          score: 100,
-          max_score: 100,
-          issues: [],
-          warnings: [],
-          details: {
-            handoff_id: handoff.id,
-            validation_score: handoff.validation_score,
-            created_at: handoff.created_at
-          }
-        };
-      },
-      required: true
-    });
-
-    // LEO v4.4.2: TEST_EVIDENCE_AUTO_CAPTURE gate
-    // SD-LEO-TESTING-GOVERNANCE-001B: Auto-ingest test reports before sub-agent orchestration
-    // Evidence: story_test_mappings often empty because test evidence not captured during handoff
-    gates.push({
-      name: 'TEST_EVIDENCE_AUTO_CAPTURE',
-      validator: async (ctx) => {
-        console.log('\n🧪 TEST EVIDENCE AUTO-CAPTURE (LEO v4.4.2)');
-        console.log('-'.repeat(50));
-
-        // 1. Check SD type exemptions (same as MANDATORY_TESTING_VALIDATION)
-        const sdType = (ctx.sd?.sd_type || 'feature').toLowerCase();
-        const EXEMPT_TYPES = ['documentation', 'docs', 'infrastructure', 'orchestrator', 'qa', 'database'];
-
-        if (EXEMPT_TYPES.includes(sdType)) {
-          console.log(`   ℹ️  ${sdType} type SD - test evidence capture SKIPPED`);
-          return {
-            passed: true,
-            score: 100,
-            max_score: 100,
-            issues: [],
-            warnings: [`Test evidence capture skipped for ${sdType} type SD`],
-            details: { skipped: true, reason: `${sdType} type exempt` }
-          };
-        }
-
-        const sdId = ctx.sd?.id || ctx.sdId;
-        const fs = await import('fs');
-        const path = await import('path');
-
-        // 2. Check for fresh existing evidence (<60 min)
-        try {
-          const { checkTestEvidenceFreshness, getLatestTestEvidence } = await import('../../../lib/test-evidence-ingest.js');
-
-          const maxAgeMinutes = parseInt(process.env.LEO_TEST_EVIDENCE_MAX_AGE_MINUTES || '60');
-          const freshnessCheck = await checkTestEvidenceFreshness(sdId, maxAgeMinutes);
-
-          if (freshnessCheck?.isFresh) {
-            console.log(`   ✅ Fresh test evidence exists (${freshnessCheck.ageMinutes?.toFixed(1) || '?'}min old)`);
-            const latestEvidence = await getLatestTestEvidence(sdId);
-            console.log(`      Verdict: ${latestEvidence?.verdict || 'UNKNOWN'}`);
-            console.log(`      Pass Rate: ${latestEvidence?.pass_rate || '?'}%`);
-
-            return {
-              passed: true,
-              score: 100,
-              max_score: 100,
-              issues: [],
-              warnings: [],
-              details: {
-                source: 'existing_fresh',
-                age_minutes: freshnessCheck.ageMinutes,
-                verdict: latestEvidence?.verdict
-              }
-            };
-          }
-        } catch (freshnessErr) {
-          console.log(`   ⚠️  Could not check freshness: ${freshnessErr.message}`);
-        }
-
-        // 3. Scan for test reports in standard locations
-        const repoPath = ctx.repoPath || process.cwd();
-        const testReportPaths = [
-          path.default.join(repoPath, 'playwright-report', 'report.json'),
-          path.default.join(repoPath, 'test-results', '.last-run.json'),
-          path.default.join(repoPath, 'coverage', 'coverage-summary.json'),
-          path.default.join(repoPath, 'playwright-report', 'results.json')
-        ];
-
-        const foundReports = [];
-        for (const reportPath of testReportPaths) {
-          if (fs.default.existsSync(reportPath)) {
-            const stats = fs.default.statSync(reportPath);
-            const ageMinutes = (Date.now() - stats.mtime.getTime()) / 60000;
-            foundReports.push({ path: reportPath, ageMinutes });
-            console.log(`   📄 Found: ${path.default.basename(reportPath)} (${ageMinutes.toFixed(0)}min old)`);
-          }
-        }
-
-        if (foundReports.length === 0) {
-          console.log('   ⚠️  No test reports found in standard locations');
-          console.log('   💡 To generate test evidence:');
-          console.log('      - E2E: npx playwright test');
-          console.log('      - Unit: npm test -- --coverage');
-
-          return {
-            passed: true, // Advisory gate - don't block
-            score: 50,
-            max_score: 100,
-            issues: [],
-            warnings: ['No test reports found - MANDATORY_TESTING_VALIDATION may fail'],
-            details: { source: 'no_reports_found' }
-          };
-        }
-
-        // 4. Call ingestTestEvidence() to capture and link to user stories
-        try {
-          const { ingestTestEvidence } = await import('../../../lib/test-evidence-ingest.js');
-
-          console.log('   📥 Ingesting test evidence...');
-
-          const ingestResult = await ingestTestEvidence({
-            sdId: sdId,
-            source: 'auto_capture_gate',
-            autoLink: true,
-            reportPaths: foundReports.map(r => r.path)
-          });
-
-          if (ingestResult?.success) {
-            console.log('   ✅ Test evidence ingested successfully');
-            console.log(`      Test Run ID: ${ingestResult.testRunId || 'created'}`);
-            console.log(`      Tests: ${ingestResult.totalTests || '?'} (${ingestResult.passedTests || '?'} passed)`);
-            console.log(`      Stories Linked: ${ingestResult.storiesLinked || 0}`);
-
-            return {
-              passed: true,
-              score: 100,
-              max_score: 100,
-              issues: [],
-              warnings: [],
-              details: {
-                source: 'auto_ingested',
-                test_run_id: ingestResult.testRunId,
-                total_tests: ingestResult.totalTests,
-                passed_tests: ingestResult.passedTests,
-                stories_linked: ingestResult.storiesLinked
-              }
-            };
-          } else {
-            console.log(`   ⚠️  Ingest returned non-success: ${ingestResult?.error || 'unknown'}`);
-            return {
-              passed: true, // Advisory - don't block
-              score: 60,
-              max_score: 100,
-              issues: [],
-              warnings: [`Test evidence ingest incomplete: ${ingestResult?.error || 'unknown'}`],
-              details: { source: 'ingest_incomplete', error: ingestResult?.error }
-            };
-          }
-        } catch (ingestErr) {
-          console.log(`   ⚠️  Auto-ingest failed: ${ingestErr.message}`);
-          console.log('   💡 Manual capture: node scripts/test-evidence-ingest.js --sd-id ' + sdId);
-
-          return {
-            passed: true, // Advisory gate - don't block
-            score: 40,
-            max_score: 100,
-            issues: [],
-            warnings: [`Test evidence auto-capture failed: ${ingestErr.message}`],
-            details: { source: 'ingest_error', error: ingestErr.message }
-          };
-        }
-      },
-      required: false // Advisory gate - MANDATORY_TESTING_VALIDATION is the blocker
-    });
-
-    // Sub-Agent Orchestration
-    gates.push({
-      name: 'SUB_AGENT_ORCHESTRATION',
-      validator: async (ctx) => {
-        console.log('\n🤖 Step 0: Sub-Agent Orchestration (PLAN_VERIFY phase)');
-        console.log('-'.repeat(50));
-
-        // SD-TYPE-AWARE SUB-AGENT EXEMPTIONS (2025-12-27)
-        // FIX 6: Query database instead of hardcoded list (2026-01-01)
-        // Exemptions defined in sd_type_validation_profiles.requires_sub_agents
-        const sdType = (ctx.sd?.sd_type || '').toLowerCase();
-
-        // Query database for SD type validation profile
-        const { data: validationProfile } = await this.supabase
-          .from('sd_type_validation_profiles')
-          .select('requires_sub_agents, validation_requirements')
-          .eq('sd_type', sdType)
-          .single();
-
-        // Check if sub-agents are NOT required for this SD type
-        const skipSubAgents = validationProfile?.requires_sub_agents === false;
-
-        if (skipSubAgents) {
-          console.log(`   ℹ️  ${sdType} type SD - sub-agent orchestration SKIPPED`);
-          console.log('   → Database: sd_type_validation_profiles.requires_sub_agents = false');
-          if (sdType === 'orchestrator') {
-            console.log('   → Orchestrator SDs: children handle sub-agent validation');
-          } else if (['documentation', 'docs'].includes(sdType)) {
-            console.log('   → Documentation SDs: no code paths to validate');
-          } else if (sdType === 'infrastructure') {
-            console.log('   → Infrastructure SDs: infrastructure-specific validation only');
-          }
-          ctx._orchestrationResult = { can_proceed: true, passed: 0, total_agents: 0, skipped: true };
-          return {
-            passed: true,
-            score: 100,
-            max_score: 100,
-            issues: [],
-            warnings: [`Sub-agent orchestration skipped for ${sdType} type SD (db: requires_sub_agents=false)`],
-            details: { skipped: true, reason: `${sdType} type - requires_sub_agents: false`, source: 'database' }
-          };
-        }
-
-        // EXEC-TO-PLAN validates completed work, so use retrospective mode
-        // This allows TESTING to use CONDITIONAL_PASS when evidence exists
-        // security_baseline: Query from sd_baseline_issues table for known pre-existing issues
-        const securityBaseline = await this._getSecurityBaseline();
-
-        // SD-QUALITY-UI-001 FIX: Check for existing PASS TESTING result before orchestration
-        // The orchestration always re-runs sub-agents which may fail in auth-constrained environments
-        // For retrospective mode, accept existing PASS results as sufficient evidence
-        const { data: existingTestingPass } = await this.supabase
-          .from('sub_agent_execution_results')
-          .select('verdict, created_at, detailed_analysis')
-          .eq('sd_id', ctx.sdId)
-          .eq('sub_agent_code', 'TESTING')
-          .eq('verdict', 'PASS')
-          .order('created_at', { ascending: false })
-          .limit(1);
-
-        // If we have a recent PASS result, skip orchestration for TESTING
-        if (existingTestingPass && existingTestingPass.length > 0) {
-          const passAge = (Date.now() - new Date(existingTestingPass[0].created_at)) / 3600000;
-          if (passAge < 24) { // Accept PASS results from last 24 hours
-            console.log(`   ✅ Found existing TESTING PASS result (${passAge.toFixed(1)}h old) - using cached result`);
-            ctx._orchestrationResult = {
-              can_proceed: true,
-              verdict: 'PASS',
-              passed: 1,
-              total_agents: 1,
-              message: 'TESTING already passed - using cached result',
-              results: [{ sub_agent_code: 'TESTING', verdict: 'PASS' }]
-            };
-            return {
-              passed: true,
-              score: 100,
-              max_score: 100,
-              issues: [],
-              warnings: ['Using cached TESTING PASS result'],
-              details: ctx._orchestrationResult
-            };
-          }
-        }
-
-        const result = await orchestrate('PLAN_VERIFY', ctx.sdId, {
-          validation_mode: 'retrospective',
-          security_baseline: securityBaseline
-        });
-        ctx._orchestrationResult = result;
-
-        if (!result.can_proceed) {
-          return {
-            passed: false,
-            score: 0,
-            max_score: 100,
-            issues: [result.message, `Failed agents: ${result.failed}`, `Blocked agents: ${result.blocked}`],
-            warnings: []
-          };
-        }
-
-        console.log(`✅ Sub-agent orchestration passed: ${result.passed}/${result.total_agents} agents`);
-
-        return {
-          passed: true,
-          score: 100,
-          max_score: 100,
-          issues: [],
-          warnings: [],
-          details: result
-        };
-      },
-      required: true
-    });
-
-    // LEO v4.4.2: MANDATORY_TESTING_VALIDATION gate
-    // SD-LEO-TESTING-GOVERNANCE-001A: Enforce TESTING sub-agent execution
-    // Evidence: 14.6% of SDs completed without TESTING validation
-    gates.push({
-      name: 'MANDATORY_TESTING_VALIDATION',
-      validator: async (ctx) => {
-        console.log('\n🧪 MANDATORY TESTING VALIDATION (LEO v4.4.2)');
-        console.log('-'.repeat(50));
-
-        // 1. Check SD type exemptions
-        // SD-LEARN-010:US-001: feature and qa SDs REQUIRE TESTING validation
-        // Only documentation, infrastructure, orchestrator, database types are exempt
-        const sdType = (ctx.sd?.sd_type || 'feature').toLowerCase();
-        const EXEMPT_TYPES = ['documentation', 'docs', 'infrastructure', 'orchestrator', 'database'];
-
-        if (EXEMPT_TYPES.includes(sdType)) {
-          console.log(`   ℹ️  ${sdType} type SD - TESTING validation SKIPPED`);
-          console.log('   → No code paths to validate');
-          return {
-            passed: true,
-            score: 100,
-            max_score: 100,
-            issues: [],
-            warnings: [`TESTING skipped for ${sdType} type SD`],
-            details: { skipped: true, reason: sdType }
-          };
-        }
-
-        // 2. Query for TESTING sub-agent execution
-        const sdUuid = ctx.sd?.id || ctx.sdId;
-        const { data: testingResults, error } = await this.supabase
-          .from('sub_agent_execution_results')
-          .select('id, verdict, confidence, created_at')
-          .eq('sd_id', sdUuid)
-          .eq('sub_agent_code', 'TESTING')
-          .order('created_at', { ascending: false })
-          .limit(1);
-
-        if (error) {
-          console.log(`   ⚠️  Error checking TESTING execution: ${error.message}`);
-          return {
-            passed: false,
-            score: 0,
-            max_score: 100,
-            issues: [`Failed to verify TESTING execution: ${error.message}`],
-            warnings: []
-          };
-        }
-
-        // 3. Validate execution exists
-        // SD-LEARN-010:US-001: ERR_TESTING_REQUIRED error code for missing TESTING execution
-        if (!testingResults?.length) {
-          console.log('   ❌ ERR_TESTING_REQUIRED: TESTING sub-agent must complete before EXEC-TO-PLAN');
-          console.log('\n   REMEDIATION:');
-          console.log('   1. Run TESTING sub-agent before completing EXEC phase');
-          console.log('   2. Command: node scripts/orchestrate-phase-subagents.js PLAN_VERIFY ' + (ctx.sdId || sdUuid));
-          console.log('   3. Ensure all E2E tests pass');
-          console.log('   4. Re-run EXEC-TO-PLAN handoff');
-          return {
-            passed: false,
-            score: 0,
-            max_score: 100,
-            issues: ['ERR_TESTING_REQUIRED: TESTING sub-agent must complete before EXEC-TO-PLAN for feature/qa SDs'],
-            warnings: []
-          };
-        }
-
-        // 4. Validate verdict is acceptable
-        const result = testingResults[0];
-        console.log(`   📊 TESTING result found: ${result.verdict} (${result.confidence}% confidence)`);
-
-        if (!['PASS', 'CONDITIONAL_PASS'].includes(result.verdict)) {
-          console.log(`   ❌ TESTING verdict ${result.verdict} - must pass`);
-          return {
-            passed: false,
-            score: 0,
-            max_score: 100,
-            issues: [`TESTING verdict ${result.verdict} - must be PASS or CONDITIONAL_PASS`],
-            warnings: []
-          };
-        }
-
-        // 5. Validate freshness (default 24h)
-        const maxAgeHours = parseInt(process.env.LEO_TESTING_MAX_AGE_HOURS || '24');
-        const ageHours = (Date.now() - new Date(result.created_at)) / 3600000;
-
-        if (ageHours > maxAgeHours) {
-          console.log(`   ⚠️  TESTING results stale (${ageHours.toFixed(1)}h old, max ${maxAgeHours}h)`);
-          return {
-            passed: false,
-            score: 50,
-            max_score: 100,
-            issues: [`TESTING results stale (${ageHours.toFixed(1)}h old, max ${maxAgeHours}h)`],
-            warnings: []
-          };
-        }
-
-        console.log('   ✅ TESTING validation passed');
-        console.log(`      Verdict: ${result.verdict}`);
-        console.log(`      Age: ${ageHours.toFixed(1)}h (max ${maxAgeHours}h)`);
-
-        return {
-          passed: true,
-          score: 100,
-          max_score: 100,
-          issues: [],
-          warnings: [],
-          details: {
-            verdict: result.verdict,
-            confidence: result.confidence,
-            age_hours: ageHours.toFixed(1),
-            max_age_hours: maxAgeHours
-          }
-        };
-      },
-      required: true
-    });
-
-    // BMAD Validation
-    gates.push({
-      name: 'BMAD_EXEC_TO_PLAN',
-      validator: async (ctx) => {
-        const result = await validateBMADForExecToPlan(ctx.sdId, this.supabase);
-        ctx._bmadResult = result;
-        return result;
-      },
-      required: true
-    });
-
-    // Gate 2: Implementation Fidelity
-    gates.push({
-      name: 'GATE2_IMPLEMENTATION_FIDELITY',
-      validator: async (ctx) => {
-        console.log('\n🚪 GATE 2: Implementation Fidelity Validation');
-        console.log('-'.repeat(50));
-        return validateGate2ExecToPlan(ctx.sdId, this.supabase);
-      },
-      required: true
-    });
-
-    // RCA Gate
-    gates.push({
-      name: 'RCA_GATE',
-      validator: async (ctx) => {
-        console.log('\n🔍 Step 1: RCA Gate Validation');
-        console.log('-'.repeat(50));
-        return this._validateRCAGate(ctx.sdId);
-      },
-      required: true
-    });
-
-    // LEO v4.4.0: Human Verification Gate
-    // Validates that feature SDs have human-verifiable outcomes (smoke tests, LLM UX)
-    gates.push({
-      name: 'HUMAN_VERIFICATION_GATE',
-      validator: async (ctx) => {
-        console.log('\n👤 Human Verification Gate (LEO v4.4.0)');
-        console.log('-'.repeat(50));
-
-        // Load the human verification validator dynamically
-        const { validateHumanVerification } = await import('../../human-verification-validator.js');
-
-        const result = await validateHumanVerification(ctx.sd?.id || ctx.sdId);
-
-        if (result.skipped) {
-          console.log(`   ℹ️  Human verification skipped: ${result.reason}`);
-          return {
-            passed: true,
-            score: 100,
-            max_score: 100,
-            issues: [],
-            warnings: [`Human verification skipped for sd_type: ${result.sdType || 'unknown'}`],
-            details: result
-          };
-        }
-
-        if (result.passed) {
-          console.log('   ✅ Human verification passed');
-          if (result.llmUxScore) {
-            console.log(`      LLM UX Score: ${result.llmUxScore}/100`);
-          }
-          if (result.smokeTestStepsCount) {
-            console.log(`      Smoke test steps: ${result.smokeTestStepsCount}`);
-          }
-          return {
-            passed: true,
-            score: 100,
-            max_score: 100,
-            issues: [],
-            warnings: [],
-            details: result
-          };
-        }
-
-        // Failed - provide detailed issues
-        console.log(`   ❌ Human verification failed: ${result.reason}`);
-        const issues = result.issues?.map(i => i.message || i) || [result.reason];
-        const actionRequired = result.issues?.find(i => i.actionRequired)?.actionRequired;
-
-        if (actionRequired) {
-          console.log(`\n   ACTION REQUIRED: ${actionRequired}`);
-        }
-
-        return {
-          passed: false,
-          score: 0,
-          max_score: 100,
-          issues,
-          warnings: [],
-          details: result,
-          remediation: actionRequired
-        };
-      },
-      // LEO v4.4.1: Human verification now REQUIRED for feature/api SDs
-      // ROOT CAUSE FIX: smoke_test_steps had 50% adherence because gate was advisory-only
-      // Now blocking for user-facing SDs to ensure human verification actually happens
-      required: true,
-      // Dynamic skip for non-feature SDs (determined by validator)
-      advisory: false
-    });
-
-    return gates;
-  }
-
-  async executeSpecific(sdId, sd, options, gateResults) {
-    // Load PRD
-    // SD ID Schema Cleanup: Use sd.id directly (uuid_id deprecated)
-    const prd = await this.prdRepo?.getBySdId(sd.id);
-
-    if (!prd) {
-      console.warn('⚠️  No PRD found for SD');
-    }
-
-    // LEO v4.3.4: Unified Test Evidence Validation
-    let testEvidenceResult = null;
-    console.log('\n🧪 Step 2: Unified Test Evidence Validation (LEO v4.3.4)');
-    console.log('-'.repeat(50));
-
-    // SD-TYPE-AWARE E2E EXEMPTIONS (2025-12-27)
-    // Based on retrospective analysis: 21% of action items were inappropriate for SD type
-    const sdType = (sd?.sd_type || '').toLowerCase();
-    const EXEMPT_FROM_E2E = ['orchestrator', 'documentation', 'docs'];
-    const E2E_OPTIONAL = ['infrastructure'];
-
-    if (EXEMPT_FROM_E2E.includes(sdType)) {
-      console.log(`   ℹ️  ${sdType} type SD - E2E test validation SKIPPED`);
-      console.log(`   → Reason: ${sdType === 'orchestrator' ? 'Children handle testing' : 'No code to test'}`);
-      testEvidenceResult = {
-        skipped: true,
-        reason: `${sdType} type SD - exempt from E2E testing`,
-        total_stories: 0,
-        passing_count: 0,
-        all_passing: true
-      };
-    } else if (E2E_OPTIONAL.includes(sdType)) {
-      console.log(`   ℹ️  ${sdType} type SD - E2E testing is OPTIONAL`);
-      console.log('   → Unit tests may suffice for infrastructure changes');
-    }
-
-    if (!testEvidenceResult?.skipped) try {
-      // Query v_story_test_coverage view for comprehensive test evidence
-      testEvidenceResult = await getStoryTestCoverage(sdId);
-
-      if (testEvidenceResult.total_stories === 0) {
-        console.log('   ℹ️  No user stories to validate');
-      } else if (testEvidenceResult.all_passing) {
-        console.log(`   ✅ All ${testEvidenceResult.passing_count}/${testEvidenceResult.total_stories} stories have passing tests`);
-        console.log(`   📊 Latest test run: ${testEvidenceResult.latest_run_at || 'N/A'}`);
-      } else {
-        console.log(`   ⚠️  Test coverage: ${testEvidenceResult.passing_count}/${testEvidenceResult.total_stories} stories passing`);
-        if (testEvidenceResult.failing_stories?.length > 0) {
-          console.log('   ❌ Failing stories:');
-          testEvidenceResult.failing_stories.slice(0, 5).forEach(story => {
-            console.log(`      - ${story.story_key}: ${story.latest_test_status || 'No test evidence'}`);
-          });
-        }
-      }
-    } catch (error) {
-      console.log(`   ⚠️  Test evidence query error: ${error.message}`);
-      console.log('   → Falling back to legacy E2E mapping');
-
-      // Fallback to legacy E2E mapping if unified schema not available
-      if (prd) {
-        try {
-          const { data: userStories } = await this.supabase
-            .from('user_stories')
-            .select('id, story_id, title, status')
-            .eq('prd_id', prd.id);
-
-          if (userStories && userStories.length > 0) {
-            const e2eMapping = await mapE2ETestsToUserStories(sdId, this.supabase);
-            const coverageResult = await validateE2ECoverage(sdId, this.supabase);
-
-            if (!coverageResult.passed) {
-              console.log(`   ⚠️  E2E coverage: ${coverageResult.mapped_count}/${coverageResult.total_stories} stories mapped`);
-            } else {
-              console.log(`   ✅ E2E test mapping complete: ${coverageResult.mapped_count} stories covered`);
-            }
-            testEvidenceResult = { legacy: true, e2eMapping, coverageResult };
-          }
-        } catch (legacyError) {
-          console.log(`   ⚠️  Legacy E2E mapping error: ${legacyError.message}`);
-        }
-      }
-    }
-
-    // Auto-validate user stories
-    console.log('\n📋 Step 3: Auto-Validate User Stories');
-    console.log('-'.repeat(50));
-
-    try {
-      const validationResult = await autoValidateUserStories(sdId, this.supabase);
-      if (validationResult.success) {
-        console.log(`   ✅ Validated ${validationResult.validated_count} user stories`);
-      }
-    } catch (error) {
-      console.log(`   ⚠️  User story validation error: ${error.message}`);
-    }
-
-    // Auto-complete deliverables
-    console.log('\n📦 Step 4: Auto-Complete Deliverables Verification');
-    console.log('-'.repeat(50));
-
-    let deliverablesStatus = null;
-    try {
-      const needsCompletion = await checkDeliverablesNeedCompletion(sdId, this.supabase);
-      if (needsCompletion.needs_completion) {
-        const completeResult = await autoCompleteDeliverables(sdId, this.supabase);
-        deliverablesStatus = completeResult;
-        console.log(`   ✅ Auto-completed ${completeResult.completed_count || 0} deliverables`);
-      } else {
-        console.log('   ℹ️  Deliverables already complete or verified by database trigger');
-      }
-    } catch (error) {
-      console.log(`   ⚠️  Deliverables completion error: ${error.message}`);
-    }
-
-    // AI Quality Assessment (Russian Judge) - Retrospective Quality
-    const russianJudgeEnabled = process.env.RUSSIAN_JUDGE_ENABLED === 'true';
-    if (russianJudgeEnabled) {
-      try {
-        console.log('\n🤖 AI QUALITY ASSESSMENT (Russian Judge)');
-        console.log('-'.repeat(50));
-
-        // Fetch retrospective for this SD
-        const { data: retrospective } = await this.supabase
-          .from('retrospectives')
-          .select('*')
-          .eq('sd_id', sdId)
-          .order('created_at', { ascending: false })
-          .limit(1)
-          .single();
-
-        if (retrospective) {
-          const { RetrospectiveQualityRubric } = await import('../../rubrics/retrospective-quality-rubric.js');
-          const rubric = new RetrospectiveQualityRubric();
-          const retroAssessment = await rubric.validateRetrospectiveQuality(retrospective, sd);
-
-          console.log(`   Retrospective Score: ${retroAssessment.score}% (threshold: 70%)`);
-          console.log(`   Status: ${retroAssessment.passed ? 'PASSED' : 'NEEDS IMPROVEMENT'}`);
-
-          if (retroAssessment.issues && retroAssessment.issues.length > 0) {
-            console.log('\n   ⚡ Retrospective Issues:');
-            retroAssessment.issues.forEach(issue => console.log(`     - ${issue}`));
-          }
-
-          if (retroAssessment.warnings && retroAssessment.warnings.length > 0) {
-            console.log('\n   💡 Recommendations:');
-            retroAssessment.warnings.forEach(warning => console.log(`     - ${warning}`));
-          }
-
-          // Mode: ADVISORY for EXEC-TO-PLAN (log but don't block)
-          if (!retroAssessment.passed) {
-            console.log('\n   ⚠️  Note: Proceeding despite quality concerns (ADVISORY mode)');
-          } else {
-            console.log('\n   ✅ Quality assessment passed');
-          }
-
-          console.log('');
-        } else {
-          console.log('   ℹ️  No retrospective found for assessment');
-          console.log('');
-        }
-      } catch (error) {
-        console.log(`\n   ⚠️  Russian Judge unavailable: ${error.message}`);
-        console.log('   Proceeding with traditional validation only\n');
-      }
-    }
-
-    // Git commit verification
-    console.log('📝 Step 5: Git Commit Verification');
-    console.log('-'.repeat(50));
-
-    let commitVerification = null;
-
-    // SD-TYPE-AWARE GIT COMMIT EXEMPTIONS (2025-12-27)
-    // Documentation SDs may have no code commits (just markdown)
-    const GIT_COMMIT_OPTIONAL = ['documentation', 'docs'];
-
-    if (GIT_COMMIT_OPTIONAL.includes(sdType)) {
-      console.log(`   ℹ️  ${sdType} type SD - Git commit check is OPTIONAL`);
-      console.log('   → May only have markdown file changes');
-      commitVerification = { verdict: 'PASS', commit_count: 0, optional: true };
-    }
-
-    if (!commitVerification?.optional) try {
-      const { default: GitCommitVerifier } = await import('../../../verify-git-commit-status.js');
-      const appPath = this.determineTargetRepository(sd);
-      // SD-VENTURE-STAGE0-UI-001: Pass legacy_id for commit search
-      const verifier = new GitCommitVerifier(sdId, appPath, { legacyId: sd?.legacy_id });
-      commitVerification = await verifier.verify();
-
-      if (commitVerification.verdict === 'PASS') {
-        console.log('   ✅ All changes committed');
-        console.log(`   Commits: ${commitVerification.commit_count}`);
-      } else {
-        console.log('   ⚠️  Uncommitted changes detected');
-      }
-    } catch (error) {
-      console.log(`   ⚠️  Git verification error: ${error.message}`);
-    }
-
-    // Build success result
-    const orchestrationResult = gateResults.gateResults.SUB_AGENT_ORCHESTRATION?.details || {};
-    const bmadResult = gateResults.gateResults.BMAD_EXEC_TO_PLAN || {};
-
-    // STATE TRANSITION: Update user stories and PRD status
-    // Root cause fix: Handoffs should act as state machine transitions, not just validation gates
-    // 5 Whys Analysis: See SD-QA-STAGES-21-25-001 retrospective
-    console.log('\n📊 Step 6: STATE TRANSITIONS');
-    console.log('-'.repeat(50));
-
-    // 6a. Update user stories to validated/completed status
-    await this._transitionUserStoriesToValidated(sdId);
-
-    // 6b. Update PRD status to verification
-    // SD ID Schema Cleanup: Use sd.id directly (uuid_id deprecated)
-    const prdForTransition = await this.prdRepo?.getBySdId(sd.id);
-    await this._transitionPrdToVerification(prdForTransition);
-
-    // 6c. Update SD status (may fail due to progress trigger - that's expected)
-    // SD-LEO-ID-NORMALIZE-001: Normalize ID before update
-    console.log('\n   Updating SD status...');
-    try {
-      const sdCanonicalId = await normalizeSDId(this.supabase, sdId);
-      if (!sdCanonicalId) {
-        console.warn(`   ⚠️  Could not normalize SD ID: ${sdId}`);
-      } else {
-        if (sdId !== sdCanonicalId) {
-          console.log(`   ℹ️  ID normalized: "${sdId}" -> "${sdCanonicalId}"`);
-        }
-
-        const { data: updateResult, error: updateError } = await this.supabase
-          .from('strategic_directives_v2')
-          .update({
-            current_phase: 'EXEC_COMPLETE',
-            updated_at: new Date().toISOString()
-          })
-          .eq('id', sdCanonicalId)
-          .select('id')
-          .single();
-
-        if (updateError) {
-          console.warn(`   ⚠️  SD phase update note: ${updateError.message}`);
-          console.log('   ℹ️  SD completion requires PLAN-TO-LEAD handoff');
-        } else if (!updateResult) {
-          // SD-LEO-ID-NORMALIZE-001: Detect silent failures
-          console.warn('   ⚠️  SD update returned no data - possible silent failure');
-        } else {
-          console.log('   ✅ SD phase updated to EXEC_COMPLETE');
-        }
-      }
-    } catch (error) {
-      console.warn(`   ⚠️  SD update error: ${error.message}`);
-    }
-
-    // =========================================================================
-    // AUTOMATED SHIPPING: PR Creation Decision (LEO Protocol v4.3.5)
-    // =========================================================================
-    let shippingResult = null;
-    try {
-      console.log('\n🚢 [AUTO-SHIP] PR Creation Decision');
-      console.log('-'.repeat(50));
-
-      const { runAutomatedShipping } = await import('../../shipping/index.js');
-      const repoPath = this.determineTargetRepository(sd);
-
-      shippingResult = await runAutomatedShipping(
-        sdId,
-        repoPath,
-        'EXEC-TO-PLAN',
-        'PR_CREATION'
-      );
-
-      if (shippingResult.executionResult?.success) {
-        console.log(`\n   ✅ PR Created: ${shippingResult.executionResult.prUrl}`);
-      } else if (shippingResult.shouldEscalate) {
-        console.log('\n   ⚠️  PR creation escalated to human - run /ship manually');
-      } else if (shippingResult.executionResult?.deferred) {
-        console.log('\n   ⏸️  PR creation deferred - fix issues first');
-      }
-    } catch (shippingError) {
-      console.warn(`   ⚠️  Auto-shipping error (non-blocking): ${shippingError.message}`);
-      // Non-blocking - handoff still succeeds even if shipping fails
-    }
-
-    return {
-      success: true,
-      subAgents: {
-        total: orchestrationResult.total_agents,
-        passed: orchestrationResult.passed
-      },
-      bmad_validation: bmadResult,
-      test_evidence: testEvidenceResult, // LEO v4.3.4: Unified test evidence
-      deliverables: deliverablesStatus,
-      commit_verification: commitVerification,
-      // LEO v4.3.5: Automated shipping result
-      automated_shipping: shippingResult ? {
-        decision: shippingResult.decision,
-        confidence: shippingResult.confidence,
-        pr_url: shippingResult.executionResult?.prUrl,
-        pr_number: shippingResult.executionResult?.prNumber,
-        escalated: shippingResult.shouldEscalate
-      } : null,
-      // Use normalized score (weighted average 0-100%) instead of summed totalScore
-      qualityScore: gateResults.normalizedScore ?? Math.round((gateResults.totalScore / gateResults.totalMaxScore) * 100)
-    };
-  }
-
-  async _validateRCAGate(sdId) {
-    try {
-      const { data: openRCRs } = await this.supabase
-        .from('root_cause_analyses')
-        .select('id, priority, capa_status')
-        .eq('sd_id', sdId)
-        .in('priority', ['P0', 'P1'])
-        .neq('capa_status', 'verified');
-
-      if (openRCRs && openRCRs.length > 0) {
-        return {
-          passed: false,
-          score: 0,
-          max_score: 100,
-          issues: [`${openRCRs.length} P0/P1 RCRs without verified CAPAs`],
-          warnings: [],
-          gate_status: 'BLOCKED',
-          open_rcr_count: openRCRs.length,
-          blocking_rcr_ids: openRCRs.map(r => r.id)
-        };
-      }
-
-      console.log('✅ RCA gate passed');
-      return {
-        passed: true,
-        score: 100,
-        max_score: 100,
-        issues: [],
-        warnings: [],
-        gate_status: 'PASS'
-      };
-    } catch (error) {
-      // Table might not exist
-      console.log(`   ℹ️  RCA gate check skipped: ${error.message || 'table may not exist'}`);
-      return {
-        passed: true,
-        score: 100,
-        max_score: 100,
-        issues: [],
-        warnings: ['RCA table check skipped']
-      };
-    }
-  }
-
-  /**
-   * STATE TRANSITION: Update user stories to validated status
-   *
-   * Root cause fix: User stories weren't being marked as validated after implementation,
-   * causing the progress trigger to block SD completion.
-   *
-   * Updates:
-   * - validation_status = 'validated'
-   * - e2e_test_status = 'passing' (if e2e_test_path exists)
-   * - status = 'completed'
-   */
-  async _transitionUserStoriesToValidated(sdId) {
-    console.log('\n   Updating user stories...');
-
-    try {
-      // Get all user stories for this SD
-      const { data: stories, error: fetchError } = await this.supabase
-        .from('user_stories')
-        .select('id, title, e2e_test_path, status, validation_status')
-        .eq('sd_id', sdId);
-
-      if (fetchError) {
-        console.log(`   ⚠️  Could not fetch user stories: ${fetchError.message}`);
-        return;
-      }
-
-      if (!stories || stories.length === 0) {
-        console.log('   ℹ️  No user stories to transition');
-        return;
-      }
-
-      // Update each story
-      let updatedCount = 0;
-      for (const story of stories) {
-        const updates = {
-          status: 'completed',
-          validation_status: 'validated',
-          updated_at: new Date().toISOString()
-        };
-
-        // Only set e2e_test_status if test path exists
-        if (story.e2e_test_path) {
-          updates.e2e_test_status = 'passing';
-        }
-
-        const { error: updateError } = await this.supabase
-          .from('user_stories')
-          .update(updates)
-          .eq('id', story.id);
-
-        if (!updateError) {
-          updatedCount++;
-        }
-      }
-
-      console.log(`   ✅ ${updatedCount}/${stories.length} user stories transitioned to validated/completed`);
-    } catch (error) {
-      console.log(`   ⚠️  User story transition error: ${error.message}`);
-    }
-  }
-
-  /**
-   * STATE TRANSITION: Update PRD status to verification
-   *
-   * Root cause fix: PRD status wasn't being updated after EXEC, causing PLAN-TO-LEAD
-   * to fail with "PRD status is 'in_progress', expected 'verification' or 'completed'"
-   */
-  async _transitionPrdToVerification(prd) {
-    if (!prd) {
-      console.log('\n   ⚠️  No PRD to transition');
-      return;
-    }
-
-    console.log('\n   Updating PRD status...');
-
-    try {
-      const { error } = await this.supabase
-        .from('product_requirements_v2')
-        .update({
-          status: 'verification',
-          phase: 'verification',
-          updated_at: new Date().toISOString()
-        })
-        .eq('id', prd.id);
-
-      if (error) {
-        console.log(`   ⚠️  Could not update PRD status: ${error.message}`);
-      } else {
-        console.log('   ✅ PRD status transitioned: → verification');
-      }
-    } catch (error) {
-      console.log(`   ⚠️  PRD transition error: ${error.message}`);
-    }
-  }
-
-  /**
-   * Get security baseline from sd_baseline_issues table
-   *
-   * Queries the baseline issues table for known pre-existing security issues
-   * that should not block EXEC-TO-PLAN handoffs for unrelated SDs.
-   *
-   * @returns {Object} Baseline counts by issue type
-   */
-  async _getSecurityBaseline() {
-    const defaultBaseline = {
-      sql_concatenation: 0,
-      eval_usage: 0,
-      dangerous_html: 0
-    };
-
-    try {
-      // Query baseline issues for security category
-      const { data: issues, error } = await this.supabase
-        .from('sd_baseline_issues')
-        .select('description, metadata')
-        .eq('category', 'security')
-        .in('status', ['open', 'acknowledged', 'in_progress']);
-
-      if (error) {
-        // Table may not exist yet - return empty baseline
-        console.log(`   ℹ️  Security baseline query: ${error.message}`);
-        return defaultBaseline;
-      }
-
-      if (!issues || issues.length === 0) {
-        return defaultBaseline;
-      }
-
-      // Count issues by type based on description patterns
-      const baseline = { ...defaultBaseline };
-
-      for (const issue of issues) {
-        const desc = (issue.description || '').toLowerCase();
-        const issueType = issue.metadata?.issue_type;
-
-        if (issueType === 'sql_concatenation' || desc.includes('sql') || desc.includes('concatenat')) {
-          baseline.sql_concatenation++;
-        } else if (issueType === 'eval_usage' || desc.includes('eval')) {
-          baseline.eval_usage++;
-        } else if (issueType === 'dangerous_html' || desc.includes('innerhtml') || desc.includes('dangerous')) {
-          baseline.dangerous_html++;
-        }
-      }
-
-      console.log(`   📊 Security baseline loaded: ${issues.length} known issues`);
-      return baseline;
-    } catch (error) {
-      console.log(`   ⚠️  Security baseline error: ${error.message}`);
-      return defaultBaseline;
-    }
-  }
-
-  getRemediation(gateName) {
-    const remediations = {
-      'SUB_AGENT_ORCHESTRATION': 'Fix sub-agent failures before creating EXEC→PLAN handoff. Review sub-agent results and address issues.',
-      'BMAD_EXEC_TO_PLAN': 'Ensure all test plans are complete and E2E test coverage is 100%.',
-      'GATE2_IMPLEMENTATION_FIDELITY': [
-        'Review Gate 2 details to see which requirements were not met:',
-        '- Testing: Unit tests executed & passing (MANDATORY)',
-        '- Server restart: Dev server restarted & verified (MANDATORY)',
-        '- Code quality: No stubbed/incomplete code (MANDATORY)',
-        '- Directory: Working in correct application (MANDATORY)',
-        '- Ambiguity: All FIXME/TODO/HACK comments resolved (MANDATORY)',
-        'After fixing issues, re-run this handoff'
-      ].join('\n'),
-      'RCA_GATE': 'All P0/P1 RCRs must have verified CAPAs before handoff. Run: node scripts/root-cause-agent.js capa verify --capa-id <UUID>',
-      'MANDATORY_TESTING_VALIDATION': [
-        'ERR_TESTING_REQUIRED: TESTING sub-agent is MANDATORY for feature/qa SDs.',
-        '',
-        'STEPS TO RESOLVE:',
-        '1. Run TESTING sub-agent before completing EXEC phase',
-        '2. Command: node scripts/orchestrate-phase-subagents.js PLAN_VERIFY <SD-ID>',
-        '3. Ensure all E2E tests pass',
-        '4. Re-run EXEC-TO-PLAN handoff',
-        '',
-        'EXEMPT SD TYPES: documentation, docs, infrastructure, orchestrator, database',
-        'REQUIRED SD TYPES: feature, qa (SD-LEARN-010:US-001)'
-      ].join('\n'),
-      'TEST_EVIDENCE_AUTO_CAPTURE': [
-        'Test evidence auto-capture helps populate story_test_mappings.',
-        '',
-        'TO GENERATE TEST EVIDENCE:',
-        '1. Run E2E tests: npx playwright test',
-        '2. Run unit tests: npm test -- --coverage',
-        '3. Manual ingest: node scripts/test-evidence-ingest.js --sd-id <SD-ID>',
-        '',
-        'REPORT LOCATIONS SCANNED:',
-        '- playwright-report/report.json',
-        '- test-results/.last-run.json',
-        '- coverage/coverage-summary.json',
-        '',
-        'This gate is advisory - MANDATORY_TESTING_VALIDATION will block if no evidence.'
-      ].join('\n')
-    };
-
-    return remediations[gateName] || null;
-  }
-
-  async _loadValidators() {
-    if (!validateBMADForExecToPlan) {
-      const bmad = await import('../../bmad-validation.js');
-      validateBMADForExecToPlan = bmad.validateBMADForExecToPlan;
-    }
-
-    if (!validateGate2ExecToPlan) {
-      const gate2 = await import('../../implementation-fidelity-validation.js');
-      validateGate2ExecToPlan = gate2.validateGate2ExecToPlan;
-    }
-
-    if (!orchestrate) {
-      const orch = await import('../../../orchestrate-phase-subagents.js');
-      orchestrate = orch.orchestrate;
-    }
-
-    if (!getValidationRequirements) {
-      const sdType = await import('../../../../lib/utils/sd-type-validation.js');
-      getValidationRequirements = sdType.getValidationRequirements;
-    }
-
-    if (!mapE2ETestsToUserStories) {
-      const e2e = await import('../map-e2e-tests-to-stories.js');
-      mapE2ETestsToUserStories = e2e.mapE2ETestsToUserStories;
-      validateE2ECoverage = e2e.validateE2ECoverage;
-    }
-
-    if (!autoValidateUserStories) {
-      const validate = await import('../../../auto-validate-user-stories-on-exec-complete.js');
-      autoValidateUserStories = validate.autoValidateUserStories;
-    }
-
-    if (!autoCompleteDeliverables) {
-      const deliverables = await import('../auto-complete-deliverables.js');
-      autoCompleteDeliverables = deliverables.autoCompleteDeliverables;
-      checkDeliverablesNeedCompletion = deliverables.checkDeliverablesNeedCompletion;
-    }
-
-    // LEO v4.3.4: Unified test evidence functions
-    if (!getStoryTestCoverage) {
-      const testEvidence = await import('../../../lib/test-evidence-ingest.js');
-      getStoryTestCoverage = testEvidence.getStoryTestCoverage;
-    }
-  }
-}
-
-export default ExecToPlanExecutor;
+// Re-export everything from the modular structure
+export * from './exec-to-plan/index.js';
+export { ExecToPlanExecutor } from './exec-to-plan/index.js';
+export { default } from './exec-to-plan/index.js';
