+++ scripts/modules/handoff/executors/exec-to-plan/gates/index.js
@@ -1,0 +1,15 @@
+/**
+ * Gate Exports Index for EXEC-TO-PLAN
+ * Part of SD-LEO-REFACTOR-EXECTOPLAN-001
+ *
+ * Re-exports all gate creators for EXEC-TO-PLAN handoff
+ */
+
+export { createPrerequisiteCheckGate } from './prerequisite-check.js';
+export { createTestEvidenceAutoCaptureGate } from './test-evidence-auto-capture.js';
+export { createSubAgentOrchestrationGate } from './sub-agent-orchestration.js';
+export { createMandatoryTestingValidationGate } from './mandatory-testing-validation.js';
+export { createBMADValidationGate } from './bmad-validation.js';
+export { createGate2ImplementationFidelityGate } from './gate2-implementation-fidelity.js';
+export { createRCAGate } from './rca-gate.js';
+export { createHumanVerificationGate } from './human-verification-gate.js';
